⚡  chore: move markdown component
moved to: /ui/primitives/atoms
