+++ src/components/index.ts
@@ -3,4 +3,3 @@ export * from './canvas';
 export * from './readme-result';
 export * from './inputs';
 export * from './group-fields';
-export * from './markdown';
